# docs/swagger.yaml
basePath: /
definitions:
...
  dto.RoleMenus:
    properties:
      id_menu:
        items:
          type: integer
        type: array
      id_role:
        type: integer
    type: object
...
  dto.UserRoles:
    properties:
      id_role:
        items:
          type: integer
        type: array
      id_user:
        type: integer
    type: object
...
host: localhost:5050/api
info:
  contact: {}
  description: INSIST adalah sistem manajemen manufaktur milik PT. Indoseiki Metalutama.
    REST API Documentation untuk INSIST menjelaskan bagaimana pengembang dapat menggunakan
    API ini untuk mengelola dan mengakses data manufaktur perusahaan dengan aman dan
    efisien, menggunakan protokol HTTP dan metode REST seperti GET, POST, PUT, & DELETE.
  title: INSIST REST API Documentation
  version: "1.0"
paths:
...
  /admin/master/role-menu:
    get:
      consumes:
      - application/json
      description: Retrieves Role Menus with pagination and optional search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of rows per page
        in: query
        name: rows
        type: integer
      - description: Search keyword for filtering Role Menu
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data found successfully
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: No data found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a list of Role Menus
      tags:
      - Role Menu
    post:
      consumes:
      - application/json
      description: Create a new Role Menu with the provided details
      parameters:
      - description: Role Menu details
        in: body
        name: RoleMenu
        required: true
        schema:
          $ref: '#/definitions/dto.RoleMenus'
      produces:
      - application/json
      responses:
        "201":
          description: Role Menu created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new Role Menu
      tags:
      - Role Menu
  /admin/master/role-menu/{id}:
    delete:
      description: Delete a RoleMenu by its ID
      parameters:
      - description: Role Menu ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Role Menu deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: RoleMenu not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a roleMenu
      tags:
      - Role Menu
    get:
      consumes:
      - application/json
      description: Retrieve a specific Role Menu by its ID
      parameters:
      - description: Role Menu ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Role Menu found successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Role Menu not found'
          schema:
            additionalProperties: true
            type: object
      summary: Get Role Menu by ID
      tags:
      - Role Menu
...
  /admin/master/user-role:
    get:
      consumes:
      - application/json
      description: Retrieves User Roles with pagination and optional search
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of rows per page
        in: query
        name: rows
        type: integer
      - description: Search keyword for filtering User Role
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data found successfully
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: No data found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a list of User Roles
      tags:
      - User Role
    post:
      consumes:
      - application/json
      description: Create a new User Role with the provided details
      parameters:
      - description: User Role details
        in: body
        name: UserRole
        required: true
        schema:
          $ref: '#/definitions/dto.UserRoles'
      produces:
      - application/json
      responses:
        "201":
          description: User Role created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new User Role
      tags:
      - User Role
  /admin/master/user-role/{id}:
    delete:
      description: Delete a UserRole by its ID
      parameters:
      - description: User Role ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: User Role deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: UserRole not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a userRole
      tags:
      - User Role
    get:
      consumes:
      - application/json
      description: Retrieve a specific User Role by its ID
      parameters:
      - description: User Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User Role found successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: User Role not found'
          schema:
            additionalProperties: true
            type: object
      summary: Get User Role by ID
      tags:
      - User Role
...
swagger: "2.0"